Added all ui_*.h files.
Someone modified the path for finding those headers and put it into the
project directory. But the problem is that they are headers produced by
building the project and they weren't being produced. So, I built it in
my own directory, copied the files and put them in git directory,
modifying the other ones and adding the ui_report.h (that didn't exist
-> it was/is being found still in my directory)